linux fixes
