+++ Plane Shooter/Rysowanie_poziomu.cpp
@@ -1,30 +1,30 @@
 #include "biblioteki.h"
-#include "Edytor.cpp"
-#include "Random_Font.hpp"
-#include <fstream>
+#include "loader.hpp"
 
+ALLEGRO_BITMAP *obrazek1 = loader("bitmapy/sciana_Placeholder.png");
+ALLEGRO_BITMAP *obrazek2 = loader("bitmapy/Droga_Placeholder.png");
+ALLEGRO_BITMAP *Enemy1 = loader("Enemies/Enemy1_Placeholder.png");
+ALLEGRO_BITMAP *Enemy2 = loader("Enemies/Enemy2_Placeholder.png");
+ALLEGRO_BITMAP *Player = loader("Player/Player_Placeholder.png");
+ALLEGRO_BITMAP *bron1 = loader("bron/bron1_Placeholder.png");
+ALLEGRO_BITMAP *bron2 = loader("bron/bron2_Placeholder.png");
+ALLEGRO_BITMAP *bonus1 = loader("bonus/bonus1_Placeholder.png");
+ALLEGRO_BITMAP *bonus2 = loader("bonus/bonus2_Placeholder.png");
 
-using namespace std;
-
-const float FPS = 60;
-
-int main();
-int Edytor();
-int loader();
-
-const int SCREEN_W = 640;
-const int SCREEN_H = 480;
-const int CURSOR_SIZE = 16;
-
-int tabRP[20][15];
+ALLEGRO_BITMAP *bonus3 = loader("bonus/bonus3_Placeholder.png");
+ALLEGRO_BITMAP *bonus4 = loader("bonus/bonus4_Placeholder.png");
+ALLEGRO_BITMAP *bonus5 = loader("bonus/bonus5_Placeholder.png");
+ALLEGRO_BITMAP *bron3 = loader("bron/bron3_Placeholder.png");
+ALLEGRO_BITMAP *bron4 = loader("bron/bron4_Placeholder.png");
+ALLEGRO_BITMAP *bron5 = loader("bron/bron5_Placeholder.png");
 
 int Rysowanie_Poziomu()
 {
+    int tabRP[20][15];
 
 
-
-	fstream Hub;
-	Hub.open("Mapy/Hub.txt", ios::in | ios::out);
+	std::fstream Hub;
+	Hub.open("Mapy/Hub.txt", std::ios::in | std::ios::out);
 	for (int i = 0; i < 20; i++)
 	{
 		for (int j = 0; j < 15; j++)
@@ -32,9 +32,9 @@ int Rysowanie_Poziomu()
 			Hub >> tabRP[i][j];
 		}
 	}
-	cout << "wczytano mape" << endl;
+	std::cout << "wczytano mape" << std::endl;
 	Hub.close();
-	int test = loader();
+	//int test = loader();//error: too few arguments to function 'ALLEGRO_BITMAP* loader(std::__cxx11::string)'
 
 	for (int i = 0; i < 20; i++)
 	{
@@ -46,14 +46,13 @@ int Rysowanie_Poziomu()
 				{
 					al_draw_bitmap(obrazek1, i * 32, j * 32, 0);
 				}
-				else if (tab[i][j] == 2)
+				else if (tabRP[i][j] == 2)//tab[i][j] was compared before, probably bug.
 				{
 					al_draw_bitmap(obrazek2, i * 32, j * 32, 0);
 				}
 			}
 		}
 	}
-
 
 	return 0;
 }
