+++ Plane Shooter/loader.hpp
@@ -1,24 +1,14 @@
 #ifndef LOADER_HPP_INCLUDED
 #define LOADER_HPP_INCLUDED
-#include "biblioteki.h"
-#include "loader.cpp"
 
+ALLEGRO_BITMAP *loader(std::string sciezka);
 
-ALLEGRO_BITMAP *obrazek1 = loader("bitmapy/sciana_Placeholder.png");
-ALLEGRO_BITMAP *obrazek2 = loader("bitmapy/Droga_Placeholder.png");
-ALLEGRO_BITMAP *Enemy1 = loader("Enemies/Enemy1_Placeholder.png");
-ALLEGRO_BITMAP *Enemy2 = loader("Enemies/Enemy2_Placeholder.png");
-ALLEGRO_BITMAP *Player = loader("Player/Player_Placeholder.png");
-ALLEGRO_BITMAP *bron1 = loader("bron/bron1_Placeholder.png");
-ALLEGRO_BITMAP *bron2 = loader("bron/bron2_Placeholder.png");
-ALLEGRO_BITMAP *bonus1 = loader("bonus/bonus1_Placeholder.png");
-ALLEGRO_BITMAP *bonus2 = loader("bonus/bonus2_Placeholder.png");
+class ProperLoader
+{
+public:
+	ProperLoader();
+	~ProperLoader();
 
-ALLEGRO_BITMAP *bonus3 = loader("bonus/bonus3_Placeholder.png");
-ALLEGRO_BITMAP *bonus4 = loader("bonus/bonus4_Placeholder.png");
-ALLEGRO_BITMAP *bonus5 = loader("bonus/bonus5_Placeholder.png");
-ALLEGRO_BITMAP *bron3 = loader("bron/bron3_Placeholder.png");
-ALLEGRO_BITMAP *bron4 = loader("bron/bron4_Placeholder.png");
-ALLEGRO_BITMAP *bron5 = loader("bron/bron5_Placeholder.png");
+};
 
 #endif // RANDOM_FONT_HPP_INCLUDED
